requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
dataclasses==0.8
Django==3.2.5
djangorestframework==3.12.4
httpie==2.4.0
idna==3.2
pkg-resources==0.0.0
psycopg2==2.9.1
Pygments==2.9.0
pymongo==3.12.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.2.4
typing-extensions==3.10.0.0
urllib3==1.26.6
